route x64 local AI through the UI port
